Moved raw data processing out so we don't have to import it
